Change save prediction logic of edge prediction tasks
Originally, we save prediction results only. However, unlink node prediction
tasks, it is impossible to find the corresponding edges for the edge prediction
results. In this PR, other than storing the prediction results, we also store
the source node ids and destination node ids of the corresponding target edges.
So users can map a prediction result back to the right edge according to
its source node id and its destination node id.

+++ tests/end2end-tests/graphstorm-ec/mgpu_test.sh
@@ -128,6 +128,27 @@ then
     exit -1
 fi
 
+cnt=$(ls -l /data/gsgnn_ec/prediction/user_rating_movie/ | grep predict | wc -l)
+if test $cnt != $NUM_INFO_TRAINERS
+then
+    echo "The number of saved prediction results $cnt is not equal to the number of inferers $NUM_INFO_TRAINERS"
+    exit -1
+fi
+
+cnt=$(ls -l /data/gsgnn_ec/prediction/user_rating_movie/ | grep src_nids | wc -l)
+if test $cnt != $NUM_INFO_TRAINERS
+then
+    echo "The number of saved source node ids $cnt is not equal to the number of inferers $NUM_INFO_TRAINERS"
+    exit -1
+fi
+
+cnt=$(ls -l /data/gsgnn_ec/prediction/user_rating_movie/ | grep dst_nids | wc -l)
+if test $cnt != $NUM_INFO_TRAINERS
+then
+    echo "The number of saved dst node ids results $cnt is not equal to the number of inferers $NUM_INFO_TRAINERS"
+    exit -1
+fi
+
 rm /tmp/log.txt
 
 cd $GS_HOME/tests/end2end-tests/
